+++ Runtime/Framework/Watch/WatchCollection.cs
@@ -26,7 +26,9 @@ namespace DT.UniStart {
     public UnityAction<ReadOnlyCollection<T>> AddListener(UnityAction<ReadOnlyCollection<T>> f) => this.onChange.AddListener(f);
     public UnityAction<ReadOnlyCollection<T>> AddOnceListener(UnityAction<ReadOnlyCollection<T>> f) => this.onChange.AddOnceListener(f);
     public UnityAction<ReadOnlyCollection<T>> RemoveListener(UnityAction<ReadOnlyCollection<T>> f) => this.onChange.RemoveListener(f);
-    public UnityAction<ReadOnlyCollection<T>> RemoveOnceListener(UnityAction<ReadOnlyCollection<T>> f) => this.onChange.RemoveOnceListener(f);
+    public UnityAction<ReadOnlyCollection<T>> AddListener(UnityAction<ReadOnlyCollection<T>> f, out UnityAction<ReadOnlyCollection<T>> named) => this.onChange.AddListener(f, out named);
+    public UnityAction<ReadOnlyCollection<T>> AddOnceListener(UnityAction<ReadOnlyCollection<T>> f, out UnityAction<ReadOnlyCollection<T>> named) => this.onChange.AddOnceListener(f, out named);
+    public UnityAction<ReadOnlyCollection<T>> RemoveListener(UnityAction<ReadOnlyCollection<T>> f, out UnityAction<ReadOnlyCollection<T>> named) => this.onChange.RemoveListener(f, out named);
 
     protected new void InvokeEvent() {
       base.InvokeEvent();
@@ -96,9 +98,11 @@ namespace DT.UniStart {
     public ReadOnlyDictionary<K, V> Value => this.readOnlyDictionary.Value;
 
     public UnityAction<ReadOnlyDictionary<K, V>> AddListener(UnityAction<ReadOnlyDictionary<K, V>> f) => this.onChange.AddListener(f);
+    public UnityAction<ReadOnlyDictionary<K, V>> AddListener(UnityAction<ReadOnlyDictionary<K, V>> f, out UnityAction<ReadOnlyDictionary<K, V>> named) => this.onChange.AddListener(f, out named);
     public UnityAction<ReadOnlyDictionary<K, V>> AddOnceListener(UnityAction<ReadOnlyDictionary<K, V>> f) => this.onChange.AddOnceListener(f);
+    public UnityAction<ReadOnlyDictionary<K, V>> AddOnceListener(UnityAction<ReadOnlyDictionary<K, V>> f, out UnityAction<ReadOnlyDictionary<K, V>> named) => this.onChange.AddOnceListener(f, out named);
     public UnityAction<ReadOnlyDictionary<K, V>> RemoveListener(UnityAction<ReadOnlyDictionary<K, V>> f) => this.onChange.RemoveListener(f);
-    public UnityAction<ReadOnlyDictionary<K, V>> RemoveOnceListener(UnityAction<ReadOnlyDictionary<K, V>> f) => this.onChange.RemoveOnceListener(f);
+    public UnityAction<ReadOnlyDictionary<K, V>> RemoveListener(UnityAction<ReadOnlyDictionary<K, V>> f, out UnityAction<ReadOnlyDictionary<K, V>> named) => this.onChange.RemoveListener(f, out named);
 
     protected new void InvokeEvent() {
       base.InvokeEvent();